requirements.txt
streamdeck
Pillow
pyobjc-framework-Quartz
pyobjc-framework-Cocoa
rumps
